PROJECT FILES/requirements.txt
streamlit
fastapi
uvicorn
requests
python-dotenv
sentence-transformers
pydantic-settings
pinecone
scikit-learn
pandas
matplotlib
